refactor: Move files from /git to /cli
