+++ packages/cli/src/commands/git/remote.ts
@@ -2,7 +2,7 @@ import color from '@heroku-cli/color'
 import {Command, flags} from '@heroku-cli/command'
 import * as Heroku from '@heroku-cli/schema'
 
-import Git from '../../git'
+import Git from '../../lib/git/git'
 
 export class GitRemote extends Command {
   static description = `adds a git remote to an app repo
